move more methods to harmony
move more methods to harmony
