+++ CustomAI/RealCityCargoTruckAI.cs
@@ -138,95 +138,5 @@ namespace RealCity.CustomAI
                 info.m_buildingAI.ModifyMaterialBuffer(data.m_targetBuilding, ref instance.m_buildings.m_buffer[data.m_targetBuilding], (TransferManager.TransferReason)data.m_transferType, ref num);
             }
         }
-
-        public override void SetSource(ushort vehicleID, ref Vehicle data, ushort sourceBuilding)
-        {
-            RemoveSource(vehicleID, ref data);
-            data.m_sourceBuilding = sourceBuilding;
-            if (sourceBuilding != 0)
-            {
-                BuildingManager instance = Singleton<BuildingManager>.instance;
-                BuildingInfo info = instance.m_buildings.m_buffer[sourceBuilding].Info;
-                data.Unspawn(vehicleID);
-                Randomizer randomizer = new Randomizer(vehicleID);
-                Vector3 vector;
-                Vector3 vector2;
-                info.m_buildingAI.CalculateSpawnPosition(sourceBuilding, ref instance.m_buildings.m_buffer[sourceBuilding], ref randomizer, m_info, out vector, out vector2);
-                Quaternion rotation = Quaternion.identity;
-                Vector3 forward = vector2 - vector;
-                if (forward.sqrMagnitude > 0.01f)
-                {
-                    rotation = Quaternion.LookRotation(forward);
-                }
-                data.m_frame0 = new Vehicle.Frame(vector, rotation);
-                data.m_frame1 = data.m_frame0;
-                data.m_frame2 = data.m_frame0;
-                data.m_frame3 = data.m_frame0;
-                data.m_targetPos0 = vector;
-                data.m_targetPos0.w = 2f;
-                data.m_targetPos1 = vector2;
-                data.m_targetPos1.w = 2f;
-                data.m_targetPos2 = data.m_targetPos1;
-                data.m_targetPos3 = data.m_targetPos1;
-                if ((data.m_flags & Vehicle.Flags.TransferToTarget) != 0)
-                {
-                    int num = Mathf.Min(0, data.m_transferSize - m_cargoCapacity);
-                    info.m_buildingAI.ModifyMaterialBuffer(sourceBuilding, ref instance.m_buildings.m_buffer[sourceBuilding], (TransferManager.TransferReason)data.m_transferType, ref num);
-                    // NON-STOCK CODE START
-                    ProcessGabargeIncome(vehicleID, ref data, num);
-                    /// NON-STOCK CODE END ///
-                    num = Mathf.Max(0, -num);
-                    data.m_transferSize += (ushort)num;
-                }
-                FrameDataUpdated(vehicleID, ref data, ref data.m_frame0);
-                instance.m_buildings.m_buffer[sourceBuilding].AddOwnVehicle(vehicleID, ref data);
-                if ((instance.m_buildings.m_buffer[sourceBuilding].m_flags & Building.Flags.IncomingOutgoing) != Building.Flags.None)
-                {
-                    if ((data.m_flags & Vehicle.Flags.TransferToTarget) != 0)
-                    {
-                        data.m_flags |= Vehicle.Flags.Importing;
-                    }
-                    else if ((data.m_flags & Vehicle.Flags.TransferToSource) != 0)
-                    {
-                        data.m_flags |= Vehicle.Flags.Exporting;
-                    }
-                }
-            }
-        }
-
-        private void ProcessGabargeIncome(ushort vehicleID, ref Vehicle data, int num)
-        {
-            BuildingManager instance = Singleton<BuildingManager>.instance;
-            Building building = instance.m_buildings.m_buffer[data.m_sourceBuilding];
-            if (building.Info.m_class.m_service == ItemClass.Service.Garbage)
-            {
-                float product_value = 0f;
-                switch ((TransferManager.TransferReason)data.m_transferType)
-                {
-                    case TransferManager.TransferReason.Lumber:
-                        product_value = -num * RealCityIndustryBuildingAI.GetResourcePrice((TransferManager.TransferReason)data.m_transferType);
-                        Singleton<EconomyManager>.instance.AddResource(EconomyManager.Resource.ResourcePrice, (int)product_value, ItemClass.Service.PlayerIndustry, ItemClass.SubService.PlayerIndustryForestry, ItemClass.Level.Level1);
-                        break;
-                    case TransferManager.TransferReason.Coal:
-                        product_value = -num * RealCityIndustryBuildingAI.GetResourcePrice((TransferManager.TransferReason)data.m_transferType);
-                        Singleton<EconomyManager>.instance.AddResource(EconomyManager.Resource.ResourcePrice, (int)product_value, ItemClass.Service.PlayerIndustry, ItemClass.SubService.PlayerIndustryOre, ItemClass.Level.Level1);
-                        break;
-                    case TransferManager.TransferReason.Petrol:
-                        product_value = -num * RealCityIndustryBuildingAI.GetResourcePrice((TransferManager.TransferReason)data.m_transferType);
-                        Singleton<EconomyManager>.instance.AddResource(EconomyManager.Resource.ResourcePrice, (int)product_value, ItemClass.Service.PlayerIndustry, ItemClass.SubService.PlayerIndustryOil, ItemClass.Level.Level1);
-                        break;
-                    default: DebugLog.LogToFileOnly("Error: ProcessGabargeIncome find unknow gabarge transition" + building.Info.m_class.ToString() + "transfer reason " + data.m_transferType.ToString()); break;
-                }
-            }
-        }
-
-        private void RemoveSource(ushort vehicleID, ref Vehicle data)
-        {
-            if (data.m_sourceBuilding != 0)
-            {
-                Singleton<BuildingManager>.instance.m_buildings.m_buffer[data.m_sourceBuilding].RemoveOwnVehicle(vehicleID, ref data);
-                data.m_sourceBuilding = 0;
-            }
-        }
     }
 }
