+++ CustomAI/RealCityPlayerBuildingAI.cs
@@ -1,52 +1,19 @@
 ﻿using ColossalFramework;
+using RealCity.UI;
 using RealCity.Util;
 using System;
 
 namespace RealCity.CustomAI
 {
-    public class RealCityPlayerBuildingAI: CommonBuildingAI
+    public class RealCityPlayerBuildingAI
     {
-        public override int GetResourceRate(ushort buildingID, ref Building data, EconomyManager.Resource resource)
-        {
-            if (resource == EconomyManager.Resource.Maintenance)
-            {
-                int num = data.m_productionRate;
-                if ((data.m_flags & Building.Flags.Evacuating) != Building.Flags.None)
-                {
-                    num = 0;
-                }
-                int budget = GetBudget(buildingID, ref data);
-                int num2 = GetMaintenanceCost() / 100;
-                num2 = num * budget / 100 * num2;
-                // NON-STOCK CODE START
-                float tempNum = CaculateEmployeeOutcome(buildingID, data);
-                if (budget < 100)
-                {
-                    budget = (budget * budget + 99) / 100;
-                }
-                else if (budget > 150)
-                {
-                    budget = 125;
-                }
-                else if (budget > 100)
-                {
-                    budget -= (100 - budget) * (100 - budget) / 100;
-                }
-                num2 = (int)((num2 / MainDataStore.gameExpenseDivide) + tempNum * budget);
-                /// NON-STOCK CODE END ///
-                return -num2;
-            }
-            return base.GetResourceRate(buildingID, ref data, resource);
-        }
-
-
-        public float CaculateEmployeeOutcome(ushort buildingID, Building building)
+        public static float CaculateEmployeeOutcome(ushort buildingID, Building building)
         {
             float num1 = 0;
             Citizen.BehaviourData behaviour = default(Citizen.BehaviourData);
             int aliveWorkerCount = 0;
             int totalWorkerCount = 0;
-            GetWorkBehaviour(buildingID, ref building, ref behaviour, ref aliveWorkerCount, ref totalWorkerCount);
+            BuildingUI.GetWorkBehaviour(buildingID, ref building, ref behaviour, ref aliveWorkerCount, ref totalWorkerCount);
             int budget = Singleton<EconomyManager>.instance.GetBudget(building.Info.m_class);
             int education0Salary = Math.Max((int)((budget * MainDataStore.govermentEducation0SalaryFixed * RealCityResidentAI.ProcessSalaryLandPriceAdjust(buildingID)) / 100), MainDataStore.govermentEducation0Salary);
             int education1Salary = Math.Max((int)((budget * MainDataStore.govermentEducation1SalaryFixed * RealCityResidentAI.ProcessSalaryLandPriceAdjust(buildingID)) / 100), MainDataStore.govermentEducation1Salary);
@@ -57,32 +24,18 @@ namespace RealCity.CustomAI
             num1 += behaviour.m_educated2Count * education2Salary;
             num1 += behaviour.m_educated3Count * education3Salary;
             int allWorkCount = RealCityResidentAI.TotalWorkCount(buildingID, building, true, false);
+            if (totalWorkerCount > allWorkCount)
+            {
+                allWorkCount = RealCityResidentAI.TotalWorkCount(buildingID, building, true, true);
+            }
 
-
-            if ((aliveWorkerCount == 0) && (allWorkCount!=0))
+            if ((aliveWorkerCount == 0) && (allWorkCount != 0))
             {
                 num1 = education3Salary * allWorkCount;
             }
 
-            float idex = (totalWorkerCount != 0) ? (allWorkCount / totalWorkerCount) : 1f;
-            if (totalWorkerCount > allWorkCount)
-            {
-                allWorkCount = RealCityResidentAI.TotalWorkCount(buildingID, building, true, true);
-                idex = 1f;
-            }
+            float idex = (totalWorkerCount != 0) ? (allWorkCount / totalWorkerCount) : 1;
             return num1 * idex / 16f;
-        }
-
-        public int GetBudget(ushort buildingID, ref Building buildingData)
-        {
-            ushort eventIndex = buildingData.m_eventIndex;
-            if (eventIndex != 0)
-            {
-                EventManager instance = Singleton<EventManager>.instance;
-                EventInfo info = instance.m_events.m_buffer[eventIndex].Info;
-                return info.m_eventAI.GetBudget(eventIndex, ref instance.m_events.m_buffer[eventIndex]);
-            }
-            return Singleton<EconomyManager>.instance.GetBudget(m_info.m_class);
         }
     }
 }
