+++ Patch/CargoTruckAISetSourcePatch.cs
@@ -1,0 +1,61 @@
+﻿using ColossalFramework;
+using Harmony;
+using RealCity.CustomAI;
+using RealCity.Util;
+using System;
+using System.Reflection;
+using UnityEngine;
+
+namespace RealCity.Patch
+{
+    [HarmonyPatch]
+    public class CargoTruckAISetSourcePatch
+    {
+        public static MethodBase TargetMethod()
+        {
+            return typeof(CargoTruckAI).GetMethod("SetSource", BindingFlags.Public | BindingFlags.Instance, null, new Type[] { typeof(ushort), typeof(Vehicle).MakeByRefType(), typeof(ushort) }, null);
+        }
+        public static void Prefix(ref CargoTruckAI __instance,  ushort vehicleID, ref Vehicle data, ushort sourceBuilding)
+        {
+            data.m_sourceBuilding = sourceBuilding;
+            if (sourceBuilding != 0)
+            {
+                BuildingManager instance = Singleton<BuildingManager>.instance;
+                BuildingInfo info = instance.m_buildings.m_buffer[sourceBuilding].Info;
+                if ((data.m_flags & Vehicle.Flags.TransferToTarget) != 0)
+                {
+                    var AI = (CargoTruckAI)data.Info.m_vehicleAI;
+                    DebugLog.LogToFileOnly($"CargoTruckAISetSourcePatch: m_cargoCapacity = {__instance.m_cargoCapacity} or {AI.m_cargoCapacity}");
+                    int num = Mathf.Min(0, data.m_transferSize - __instance.m_cargoCapacity);
+                    ProcessGabargeIncome(vehicleID, ref data, num);
+                }
+            }
+        }
+
+        public static void ProcessGabargeIncome(ushort vehicleID, ref Vehicle data, int num)
+        {
+            BuildingManager instance = Singleton<BuildingManager>.instance;
+            Building building = instance.m_buildings.m_buffer[data.m_sourceBuilding];
+            if (building.Info.m_class.m_service == ItemClass.Service.Garbage)
+            {
+                float product_value = 0f;
+                switch ((TransferManager.TransferReason)data.m_transferType)
+                {
+                    case TransferManager.TransferReason.Lumber:
+                        product_value = -num * RealCityIndustryBuildingAI.GetResourcePrice((TransferManager.TransferReason)data.m_transferType);
+                        Singleton<EconomyManager>.instance.AddResource(EconomyManager.Resource.ResourcePrice, (int)product_value, ItemClass.Service.PlayerIndustry, ItemClass.SubService.PlayerIndustryForestry, ItemClass.Level.Level1);
+                        break;
+                    case TransferManager.TransferReason.Coal:
+                        product_value = -num * RealCityIndustryBuildingAI.GetResourcePrice((TransferManager.TransferReason)data.m_transferType);
+                        Singleton<EconomyManager>.instance.AddResource(EconomyManager.Resource.ResourcePrice, (int)product_value, ItemClass.Service.PlayerIndustry, ItemClass.SubService.PlayerIndustryOre, ItemClass.Level.Level1);
+                        break;
+                    case TransferManager.TransferReason.Petrol:
+                        product_value = -num * RealCityIndustryBuildingAI.GetResourcePrice((TransferManager.TransferReason)data.m_transferType);
+                        Singleton<EconomyManager>.instance.AddResource(EconomyManager.Resource.ResourcePrice, (int)product_value, ItemClass.Service.PlayerIndustry, ItemClass.SubService.PlayerIndustryOil, ItemClass.Level.Level1);
+                        break;
+                    default: DebugLog.LogToFileOnly("Error: ProcessGabargeIncome find unknow gabarge transition" + building.Info.m_class.ToString() + "transfer reason " + data.m_transferType.ToString()); break;
+                }
+            }
+        }
+    }
+}
